move unit preferences button to navbar
